auth outsourced to logic
