+++ DepotMap.Backend/DepotMap.Endpoint/Controllers/AuthController.cs
@@ -1,5 +1,7 @@
 ﻿using DepotMap.Data.Context;
 using DepotMap.Entities.Models;
+using DepotMap.Entities.Models.DTOs;
+using DepotMap.Logics.Interfaces;
 using DepotMap.Logics.Services;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -11,38 +13,21 @@ namespace DepotMap.Endpoint.Controllers
     [Route("api/[controller]")]
     public class AuthController : ControllerBase
     {
-        private readonly AppDbContext _context;
-        private readonly IPasswordHasher<User> _hasher;
-        private readonly JwtService _jwtService;
+        private readonly IAuthLogic _authLogic;
 
-        public AuthController(AppDbContext context, IPasswordHasher<User> hasher, JwtService jwtService)
+        public AuthController(IAuthLogic authLogic)
         {
-            _context = context;
-            _hasher = hasher;
-            _jwtService = jwtService;
+            _authLogic = authLogic;
         }
 
         [HttpPost("login")]
         public async Task<IActionResult> Login([FromBody] LoginDto dto)
         {
-            var user = await _context.Users
-                .FirstOrDefaultAsync(u => u.Identifier == dto.Identifier);
-
-            if (user == null)
+            var token = await _authLogic.LoginAsync(dto);
+            if (token == null)
                 return Unauthorized("Hibás azonosító vagy jelszó.");
 
-            var result = _hasher.VerifyHashedPassword(user, user.PasswordHash, dto.Password);
-            if (result == PasswordVerificationResult.Failed)
-                return Unauthorized("Hibás azonosító vagy jelszó.");
-
-            var token = _jwtService.GenerateToken(user);
             return Ok(new { token });
         }
-    }
-
-    public class LoginDto
-    {
-        public string Identifier { get; set; } = null!;
-        public string Password { get; set; } = null!;
     }
 }
