+++ DepotMap.Backend/DepotMap.Endpoint/Program.cs
@@ -1,6 +1,8 @@
 using DepotMap.Data.Context;
 using DepotMap.Data.DbSeeder;
 using DepotMap.Entities.Models;
+using DepotMap.Logics.Interfaces;
+using DepotMap.Logics.Logics;
 using DepotMap.Logics.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Identity;
@@ -43,7 +45,7 @@ builder.Services.AddCors(options =>
     });
 });
 
-
+builder.Services.AddScoped<IAuthLogic, AuthLogic>();
 builder.Services.AddScoped<DbSeeder>();
 builder.Services.AddScoped<JwtService>();
 builder.Services.AddScoped<IPasswordHasher<User>, PasswordHasher<User>>();
